renamed 2 useCases package to use_cases, and DataPersistency to data_persistency
