+++ src/main/java/chat/use_cases/ChatRoomOpenInteractor.java
@@ -1,4 +1,4 @@
-package chat.useCases;
+package chat.use_cases;
 
 import chat.control.*;
 import chat.entities.ChatRoomEnt;
