+++ src/main/java/spots/use_cases/RecommendSpots.java
@@ -1,8 +1,7 @@
-package spots.useCases;
+package spots.use_cases;
 
 
 import account_creation.Account;
-import profile.Profile;
 import spots.controllers.RecsInModel;
 import spots.controllers.RecsOutBoundary;
 import spots.controllers.RecsOutModel;
