+++ src/main/java/spots/use_cases/RecsInBoundary.java
@@ -1,4 +1,4 @@
-package spots.useCases;
+package spots.use_cases;
 
 import spots.controllers.RecsInModel;
